Consolidate Transcript

Take a tibble containing the comments from a Zoom recording transcript and return a tibble that consolidates all consecutive comments from the same speaker where the time between the end of the first comment and start of the second comment is less than max_pause_sec seconds. This function addresses an issue with the Zoom transcript where the speaker is speaking a continuous sentence, but the Zoom transcript will cut the comment into two lines. For example, a comment of "This should be a single sentence." is often split into "This should be" and "a single sentence". This function stitches those together into "This should be a single sentence." where the start time of the consolidated comment will be the beginning of the first row and the end time of the consolidated comment will be the ending of the last row.

Usage

consolidate_transcript(df = NULL, max_pause_sec = 1)

Arguments

df

A tibble containing transcript comments with columns: name, start, end, comment

max_pause_sec

Maximum pause in seconds between comments to consolidate (default: 1)

Value

A tibble with consolidated comments

transcript.

consolidate_transcript(df = "NULL")